Feature/v2.6

+++ go_sample_app/publisher/server.go
@@ -1,12 +1,12 @@
 /*
-Copyright (c) 2020 Intel Corporation.
+Copyright (c) 2021 Intel Corporation
 
-Permission is hereby granted, free of charge, to any person obtaining a copy of
-this software and associated documentation files (the "Software"), to deal in
-the Software without restriction, including without limitation the rights to
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
-of the Software, and to permit persons to whom the Software is furnished to do
-so, subject to the following conditions:
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
 
 The above copyright notice and this permission notice shall be included in all
 copies or substantial portions of the Software.
@@ -25,39 +25,45 @@ package main
 import (
 	eiicfgmgr "ConfigMgr/eiiconfigmgr"
 	eiimsgbus "EIIMessageBus/eiimsgbus"
-	"github.com/golang/glog"
 	"fmt"
+
+	"github.com/golang/glog"
 )
 
 func start_server() {
 
 	configmgr, err := eiicfgmgr.ConfigManager()
-	if(err != nil) {
+	if err != nil {
 		glog.Fatal("Config Manager initialization failed...")
 	}
 	defer configmgr.Destroy()
 
+	numOfServers, _ := configmgr.GetNumServers()
+	if numOfServers == -1 {
+		glog.Errorf("No server instances found, exiting...")
+		return
+	}
 	// serverctx, err := configmgr.GetServerByName("echo_service")
 	serverctx, err := configmgr.GetServerByIndex(0)
-	if(err != nil) {
+	if err != nil {
 		glog.Fatal("GetServerByIndex is failed")
 	}
 	defer serverctx.Destroy()
 
 	interfaceVal, err := serverctx.GetInterfaceValue("Name")
-	if(err != nil){
+	if err != nil {
 		fmt.Printf("Error to GetInterfaceValue of 'Name': %v\n", err)
 		return
 	}
 
 	serviceName, err := interfaceVal.GetString()
-	if(err != nil) {
+	if err != nil {
 		fmt.Printf("Error to GetString value of 'Name'%v\n", err)
 		return
 	}
 
 	config, err := serverctx.GetMsgbusConfig()
-	if(err != nil) {
+	if err != nil {
 		glog.Fatal("Error occured with error:%v", err)
 	}
